% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{bootstrap_se}
\alias{bootstrap_se}
\title{Bootstrap standard error for synthdid estimates}
\usage{
bootstrap_se(estimate, replications)
}
\arguments{
\item{estimate}{A \code{synthdid_estimate} object.}

\item{replications}{Number of bootstrap replications.}
}
\value{
Scalar bootstrap standard error.
}
\description{
Implements Algorithm 2 from Arkhangelsky et al. to compute a bootstrap
standard error.
}
\keyword{internal}
