% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_engine.R
\name{build_diagnostic_spec}
\alias{build_diagnostic_spec}
\title{Build diagnostic spec data frames}
\usage{
build_diagnostic_spec(spec, payload)
}
\description{
Build diagnostic spec data frames
}
\keyword{internal}
