% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_engine.R
\name{build_trajectory_spec}
\alias{build_trajectory_spec}
\title{Build trajectory spec data frames}
\usage{
build_trajectory_spec(spec, payload)
}
\description{
Build trajectory spec data frames
}
\keyword{internal}
