% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CPS}
\alias{CPS}
\title{CPS}
\format{
A data frame with 2000 rows and 8 variables.
\describe{
\item{state}{state}
\item{year}{year}
\item{log_wage}{log_wage}
\item{hours}{hours}
\item{urate}{urate}
\item{min_wage}{min_wage}
\item{open_carry}{open_carry}
\item{abort_ban}{abort_ban}
}
}
\usage{
data(CPS)
}
\description{
CPS
}
