% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdid_lib.R
\name{did_est}
\alias{did_est}
\title{Returns a diff-in-diff estimate for treatment effect with a treated block. 
Also returns, as an attribute of the return value, a jacknife estimate of its standard error.}
\usage{
did_est(Y, n_0, T_0)
}
\arguments{
\item{Y, }{the observation matrix.}

\item{n_0, }{the number of control units. Rows 1-n_0 of Y correspond to the control units.}

\item{T_0, }{the number of pre-treatment time steps. Columns 1-T_0 of Y correspond to pre-treatment time steps.}
}
\value{
An average treatment effect estimate, with a standard error estimate attached as the attribute 'se'
}
\description{
Returns a diff-in-diff estimate for treatment effect with a treated block. 
Also returns, as an attribute of the return value, a jacknife estimate of its standard error.
}
