% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_engine.R
\name{.render_pretreatment_fit}
\alias{.render_pretreatment_fit}
\title{Render the pre-treatment fit panel of a diagnostic plot}
\usage{
.render_pretreatment_fit(spec)
}
\arguments{
\item{spec}{A plot spec with pretreatment_fit data}
}
\value{
A ggplot2 object (or patchwork composite if patchwork is available)
}
\description{
Render the pre-treatment fit panel of a diagnostic plot
}
\keyword{internal}
