% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebal.R
\name{eb_estimate}
\alias{eb_estimate}
\title{Difference in differences with entropy balancing weights}
\usage{
eb_estimate(Y, N0, T0, demean = FALSE)
}
\arguments{
\item{Y}{the observation matrix.}

\item{N0}{the number of control units (N_co in the paper). Rows 1-N0 of Y correspond to the control units.}

\item{T0}{the number of pre-treatment time steps (T_pre in the paper). Columns 1-T0 of Y correspond to pre-treatment time steps.}

\item{demean}{Remove unit FEs before balancing? TRUE allows for extrapolation}
}
\value{
vector of treatment effects
}
\description{
Difference in differences with entropy balancing weights
}
