% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_engine.R
\name{extract_diagnostic_payload}
\alias{extract_diagnostic_payload}
\title{Extract diagnostic plot payload from a synthdid_estimate object}
\usage{
extract_diagnostic_payload(object)
}
\arguments{
\item{object}{A synthdid_estimate object}
}
\value{
A list with convergence trace and pre-treatment fit info
}
\description{
Extract diagnostic plot payload from a synthdid_estimate object
}
\keyword{internal}
