% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_engine.R
\name{extract_plot_payload.synthdid_estimate}
\alias{extract_plot_payload.synthdid_estimate}
\title{Extract plot payload from a synthdid_estimate object}
\usage{
extract_plot_payload.synthdid_estimate(object, include_controls = FALSE)
}
\arguments{
\item{object}{A synthdid_estimate object}

\item{include_controls}{If TRUE, include individual control trajectories}
}
\value{
A list with standardized plot data
}
\description{
Extract plot payload from a synthdid_estimate object
}
\keyword{internal}
