% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_engine.R
\name{extract_plot_payload.synthdid_staggered}
\alias{extract_plot_payload.synthdid_staggered}
\title{Extract plot payload from a synthdid_staggered object}
\usage{
extract_plot_payload.synthdid_staggered(object, type = "effect")
}
\arguments{
\item{object}{A synthdid_staggered object}

\item{type}{One of "trajectory", "weights", "effect"}
}
\value{
A list with standardized plot data
}
\description{
Extract plot payload from a synthdid_staggered object
}
\keyword{internal}
