% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{extract_vars_from_expr}
\alias{extract_vars_from_expr}
\title{Extract variable names from formula expression}
\usage{
extract_vars_from_expr(expr)
}
\arguments{
\item{expr}{An R expression from formula parsing}
}
\value{
Character vector of variable names
}
\description{
Extract variable names from formula expression
}
\keyword{internal}
