% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placebo-simulations.R
\name{fit_ar2}
\alias{fit_ar2}
\title{Estimate ar2 coefficients from iid time series}
\usage{
fit_ar2(E)
}
\arguments{
\item{E, }{a matrix with those time series as rows}
}
\value{
a vector of ar2 coefficients: c(lag-1-coefficient, lag-2-coefficient)
}
\description{
Estimate ar2 coefficients from iid time series
}
