% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{fitted.synthdid}
\alias{fitted.synthdid}
\title{Extract fitted values from synthdid object}
\usage{
\method{fitted}{synthdid}(object, ...)
}
\arguments{
\item{object}{A synthdid object}

\item{...}{Additional arguments (currently ignored)}
}
\value{
A matrix of fitted values for the control units
}
\description{
Extract fitted values from synthdid object
}
\examples{
\donttest{
data(california_prop99)
result <- synthdid(PacksPerCapita ~ treated,
  data = california_prop99,
  index = c("State", "Year")
)
fitted_vals <- fitted(result)
head(fitted_vals)
}
}
