% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format.synthdid_estimate}
\alias{format.synthdid_estimate}
\title{Format a synthdid object}
\usage{
\method{format}{synthdid_estimate}(x, ...)
}
\arguments{
\item{x}{The object to format}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A character string with formatted summary information
}
\description{
Format a synthdid object
}
\examples{
\donttest{
data(california_prop99)
setup <- panel.matrices(california_prop99)
tau.hat <- synthdid_estimate(setup$Y, setup$N0, setup$T0)
format(tau.hat)
}
}
