% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blas_threads.R
\name{is_future_parallel}
\alias{is_future_parallel}
\title{Detect if future plan is using parallel workers}
\usage{
is_future_parallel()
}
\value{
Logical indicating if parallel processing is active
}
\description{
Detect if future plan is using parallel workers
}
\keyword{internal}
