% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{jackknife}
\alias{jackknife}
\title{Jackknife standard error of function \code{theta} at samples \code{x}.}
\usage{
jackknife(x, theta)
}
\arguments{
\item{x}{vector of samples}

\item{theta}{a function which returns a scalar estimate}
}
\description{
Jackknife standard error of function \code{theta} at samples \code{x}.
}
\keyword{internal}
