% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placebo-simulations.R
\name{lindsey_density_estimate}
\alias{lindsey_density_estimate}
\title{Computes a density estimator by smoothing a histogram using Poisson regression.
Implementation of "Lindsey's method", as descrbied in Chapter 10 of
"Computer age statistical inference: algorithms, evidence, and data science'
by Bradley Efron and Trevor Hastie (2016).}
\usage{
lindsey_density_estimate(x, K, deg)
}
\arguments{
\item{x}{\itemize{
\item one-dimensional vector of data;
}}

\item{K}{\itemize{
\item number of bins in the histogram;
}}

\item{deg}{\itemize{
\item degree of natural splines used in Poisson regression;
}}
}
\value{
a list with 2 fields, centers and density, which are K-dimensional vectors containing the bin centers and estimated density within each bin respectively.
}
\description{
Computes a density estimator by smoothing a histogram using Poisson regression.
Implementation of "Lindsey's method", as descrbied in Chapter 10 of
"Computer age statistical inference: algorithms, evidence, and data science'
by Bradley Efron and Trevor Hastie (2016).
}
