% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make.panel}
\alias{make.panel}
\title{Convert a long (balanced) panel to a wide matrix}
\usage{
make.panel(panel, unit = 1, time = 2, outcome = 3, treatment = 4)
}
\arguments{
\item{panel}{A data.frame with columns consisting of units, time, outcome, and treatment indicator.}

\item{unit}{The column number/name corresponding to the unit identifier. Default is 1.}

\item{time}{The column number/name corresponding to the time identifier. Default is 2.}

\item{outcome}{The column number/name corresponding to the outcome identifier. Default is 3.}

\item{treated}{The column number/name corresponding to the treatment status. Default is 4.}
}
\value{
A list with entries \code{Y}: the data matrix, \code{N0}: the number of control units, \code{T0}:
the number of time periods before treatment.
}
\description{
Converts a data set in panel form to matrix format required by synthdid estimators.
}
\examples{
\donttest{
# Load tobacco sales in long panel format.
data("california_prop99")
# Transform to N*T matrix format required for synthdid,
# where N is the number of units and T the time periods.
setup <- make.panel(california_prop99, unit = 1, time = 2, outcome = 3, treatment = 4)

# Compute synthdid estimate
synthdid_estimate(setup$Y, setup$N0, setup$T0)
}

}
