% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{model.frame.synthdid}
\alias{model.frame.synthdid}
\title{Model frame for synthdid}
\usage{
\method{model.frame}{synthdid}(formula, ...)
}
\arguments{
\item{formula}{A synthdid object}

\item{...}{Additional arguments}
}
\description{
Model frame for synthdid
}
