% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{parse_synthdid_formula}
\alias{parse_synthdid_formula}
\title{Parse synthdid formula}
\usage{
parse_synthdid_formula(formula)
}
\arguments{
\item{formula}{A formula object}
}
\value{
A list with outcome, treatment, and covariates
}
\description{
Internal function to parse formula of the form outcome ~ treatment | covariates
}
\keyword{internal}
