% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PENN}
\alias{PENN}
\title{PENN}
\format{
A data frame with 3219 rows and 5 variables.
\describe{
\item{country}{country}
\item{year}{year}
\item{log_gdp}{log_gdp}
\item{dem}{dem}
\item{educ}{educ}
}
}
\usage{
data(PENN)
}
\description{
PENN
}
