% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{placebo_se}
\alias{placebo_se}
\title{Placebo standard error for synthdid estimates}
\usage{
placebo_se(estimate, replications)
}
\arguments{
\item{estimate}{A \code{synthdid_estimate} object.}

\item{replications}{Number of placebo replications.}
}
\value{
Scalar placebo standard error.
}
\description{
Implements Algorithm 4 from Arkhangelsky et al. by repeatedly treating random
subsets of control units as pseudo-treated to approximate the estimator's
variability.
}
\keyword{internal}
