% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdid_lib.R
\name{plot.did}
\alias{plot.did}
\title{Plots treated and average control trajectories and overlays a 2x2 diff-in-diff diagram of our estimator.
In this overlay, the treatment effect is the length of the vertical dotted line. 
Requires ggplot2}
\usage{
\method{plot}{did}(est)
}
\arguments{
\item{est, }{output of did_estimate}
}
\description{
Plots treated and average control trajectories and overlays a 2x2 diff-in-diff diagram of our estimator.
In this overlay, the treatment effect is the length of the vertical dotted line. 
Requires ggplot2
}
