% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdid_lib.R
\name{plot.synthdid}
\alias{plot.synthdid}
\title{Plots treated and synthetic control trajectories and overlays a 2x2 diff-in-diff diagram of our estimator.
In this overlay, the treatment effect is the length of the vertical dotted line. 
The weights lambda defining our synthetic pre-treatment time period are plotted below.
Requires ggplot2}
\usage{
\method{plot}{synthdid}(
  est,
  treated.name = "treated",
  control.name = "synthetic control",
  use.intercept = FALSE
)
}
\arguments{
\item{est, }{output of synthdid_estimate}

\item{treated.name, }{the name of the control curve that appears in the legend. Defaults to 'synthetic control'}
}
\description{
Plots treated and synthetic control trajectories and overlays a 2x2 diff-in-diff diagram of our estimator.
In this overlay, the treatment effect is the length of the vertical dotted line. 
The weights lambda defining our synthetic pre-treatment time period are plotted below.
Requires ggplot2
}
