% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staggered_methods.R
\name{plot_staggered_cohort}
\alias{plot_staggered_cohort}
\title{Cohort-level bar plot for staggered SDID}
\usage{
plot_staggered_cohort(object)
}
\arguments{
\item{object}{A synthdid_staggered object}
}
\value{
A ggplot2 object
}
\description{
Cohort-level bar plot for staggered SDID
}
\keyword{internal}
