% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staggered_methods.R
\name{plot_staggered_event}
\alias{plot_staggered_event}
\title{Event-study plot for staggered SDID}
\usage{
plot_staggered_event(object)
}
\arguments{
\item{object}{A synthdid_staggered object}
}
\value{
A ggplot2 object
}
\description{
Event-study plot for staggered SDID
}
\keyword{internal}
