% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_estimate.R
\name{print.panelEstimate}
\alias{print.panelEstimate}
\title{print method (prints summary table)}
\usage{
\method{print}{panelEstimate}(x)
}
\arguments{
\item{data}{\link{data.frame, matrix} data table (n x p)}
}
\description{
print method (prints summary table)
}
