% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.synthdid_estimate}
\alias{print.summary.synthdid_estimate}
\title{Print summary of synthdid object}
\usage{
\method{print}{summary.synthdid_estimate}(x, ...)
}
\arguments{
\item{x}{A summary.synthdid_estimate object}

\item{...}{Additional arguments}
}
\value{
Invisibly returns the summary object
}
\description{
Print summary of synthdid object
}
\examples{
\donttest{
data(california_prop99)
setup <- panel.matrices(california_prop99)
tau.hat <- synthdid_estimate(setup$Y, setup$N0, setup$T0)
s <- summary(tau.hat)
print(s)
}
}
