% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staggered_methods.R
\name{print.summary.synthdid_staggered}
\alias{print.summary.synthdid_staggered}
\title{Print summary of synthdid_staggered}
\usage{
\method{print}{summary.synthdid_staggered}(x, ...)
}
\arguments{
\item{x}{A summary.synthdid_staggered object}

\item{...}{Additional arguments}
}
\value{
Invisibly returns x
}
\description{
Print summary of synthdid_staggered
}
