% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{print.synthdid}
\alias{print.synthdid}
\title{Print method for synthdid objects}
\usage{
\method{print}{synthdid}(x, ...)
}
\arguments{
\item{x}{A synthdid object}

\item{...}{Additional arguments (currently ignored)}
}
\value{
Invisibly returns the original object
}
\description{
Print method for synthdid objects
}
\examples{
\donttest{
data(california_prop99)
result <- synthdid(PacksPerCapita ~ treated,
  data = california_prop99,
  index = c("State", "Year")
)
print(result)
}
}
