% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.synthdid_estimate}
\alias{print.synthdid_estimate}
\title{Print a synthdid object}
\usage{
\method{print}{synthdid_estimate}(x, ...)
}
\arguments{
\item{x}{The object to print}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisibly returns the original object
}
\description{
Print a synthdid object
}
\examples{
\donttest{
data(california_prop99)
setup <- panel.matrices(california_prop99)
tau.hat <- synthdid_estimate(setup$Y, setup$N0, setup$T0)
print(tau.hat)
}
}
