% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staggered_methods.R
\name{print.synthdid_staggered}
\alias{print.synthdid_staggered}
\title{Print method for synthdid_staggered}
\usage{
\method{print}{synthdid_staggered}(x, ...)
}
\arguments{
\item{x}{A synthdid_staggered object}

\item{...}{Additional arguments (currently ignored)}
}
\value{
Invisibly returns x
}
\description{
Print method for synthdid_staggered
}
