% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_engine.R
\name{render_effect_plot}
\alias{render_effect_plot}
\title{Render an effect plot}
\usage{
render_effect_plot(spec, subtype = "cohort")
}
\arguments{
\item{spec}{A plot spec from build_plot_spec with type="effect"}

\item{subtype}{For staggered objects: "cohort" or "event"}
}
\value{
A ggplot2 object
}
\description{
Render an effect plot
}
\keyword{internal}
