% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blas_threads.R
\name{restore_blas_threads}
\alias{restore_blas_threads}
\title{Restore BLAS thread count}
\usage{
restore_blas_threads(old)
}
\arguments{
\item{old}{Previous state from set_blas_threads()}
}
\description{
Restores BLAS threads to a previously saved state.
}
\keyword{internal}
