% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{sc_estimate}
\alias{sc_estimate}
\title{Synthetic control estimate}
\usage{
sc_estimate(Y, N0, T0, eta.omega = SYNTHDID_ETA_OMEGA_SC_DEFAULT, ...)
}
\arguments{
\item{Y}{the observation matrix.}

\item{N0}{the number of control units. Rows 1-N0 of Y correspond to the control units.}

\item{T0}{the number of pre-treatment time steps. Columns 1-T0 of Y correspond to pre-treatment time steps.}

\item{eta.omega}{determines the level of ridge regularization, zeta.omega = eta.omega * noise.level, as in synthdid_estimate.}

\item{...}{additional options for synthdid_estimate}
}
\value{
an object like that returned by synthdid_estimate
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{sc_estimate()} was deprecated in synthdid 2.0.0.
Use \code{synthdid(formula, data, method = "sc")} instead.

Takes all the same parameters as \code{\link[=synthdid_estimate]{synthdid_estimate()}}, but by default,
passes options to use the synthetic control estimator.
}
\keyword{internal}
