% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdid_lib.R
\name{sc_weight}
\alias{sc_weight}
\title{Computes regularized synthetic control weights:
gamma = argmin_g{||M g - target||_2^2 / length(target) +
                 zeta * ||g||_2^2 : sum g = 1, g >= 0}}
\usage{
sc_weight(M, target, zeta = 1)
}
\arguments{
\item{zeta.}{Defaults to 1.}
}
\value{
gamma
}
\description{
Computes regularized synthetic control weights:
gamma = argmin_g{||M g - target||_2^2 / length(target) +
                 zeta * ||g||_2^2 : sum g = 1, g >= 0}
}
