% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdid_lib.R
\name{sc_weight_FE}
\alias{sc_weight_FE}
\title{Computes regularized synthetic control weights with an intercept.
gamma = argmin_g{||M g - target - c||_2^2 / length(target) +
                      zeta * ||g||_2^2 : sum g = 1, g >= 0}
This may be useful when there are global time trends; see (2.7) in the paper.}
\usage{
sc_weight_FE(M, target, zeta = 1)
}
\arguments{
\item{zeta.}{Defaults to 1.}
}
\value{
gamma
}
\description{
Computes regularized synthetic control weights with an intercept.
gamma = argmin_g{||M g - target - c||_2^2 / length(target) +
                      zeta * ||g||_2^2 : sum g = 1, g >= 0}
This may be useful when there are global time trends; see (2.7) in the paper.
}
