% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdid_lib.R
\name{sdid_est}
\alias{sdid_est}
\title{Computes synthetic diff-in-diff estimate for an average treatment effect with a treated block. 
See Section 4.1 of the paper. Also computes a jacknife estimate of its standard error.
This uses the method described in the footnote in Section 5 of the paper if fast.var=FALSE.
If fast.var=T, we use a variant which is less computationally demanding,
in which we do not re-estimate omega in our jackknife replications.}
\usage{
sdid_est(Y, n_0, T_0, zeta = var(as.numeric(Y)), fast.var = T)
}
\arguments{
\item{Y, }{the observation matrix.}

\item{n_0, }{the number of control units. Rows 1-n_0 of Y correspond to the control units.}

\item{T_0, }{the number of pre-treatment time steps. Columns 1-T_0 of Y correspond to pre-treatment time steps.}

\item{zeta, }{the weight on an L2 penalty on the weights. See (6.1) in the paper. Defaults to var(Y).}

\item{fast.var.}{Defaults to TRUE.}
}
\value{
An average treatment effect estimate, with a standard error estimate attached as the attribute 'se'
}
\description{
Computes synthetic diff-in-diff estimate for an average treatment effect with a treated block. 
See Section 4.1 of the paper. Also computes a jacknife estimate of its standard error.
This uses the method described in the footnote in Section 5 of the paper if fast.var=FALSE.
If fast.var=T, we use a variant which is less computationally demanding,
in which we do not re-estimate omega in our jackknife replications.
}
