% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdid_lib.R
\name{sdid_predict}
\alias{sdid_predict}
\title{synthetic diff-in-diff and synthetic control predictions 
for a single missing observation, element (n, T) in an n x T matrix.}
\usage{
sdid_predict(Y, zeta = var(as.numeric(Y)))
}
\arguments{
\item{Y, }{the observation matrix}

\item{zeta, }{the weight on an L2 penalty on the weights. See (6.1) in the paper. Defaults to var(Y).}
}
\value{
a 2-vector of estimates, synthetic diff-in-diff followed by synthetic control
}
\description{
synthetic diff-in-diff and synthetic control predictions 
for a single missing observation, element (n, T) in an n x T matrix.
}
