% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sdid_rmvnorm}
\alias{sdid_rmvnorm}
\title{Simulate correlated mean zero normal variables}
\usage{
sdid_rmvnorm(n, sigma)
}
\arguments{
\item{n}{The number of samples.}

\item{sigma}{The covariance matrix.}
}
\value{
A n X ncol(sigma) matrix of draws.
}
\description{
Simulate correlated mean zero normal variables with covariance matrix
`sigma` using the Cholesky decomposition. A simple convenience
function to simulate synthetic data with.
}
\examples{
\donttest{
sigma <- cbind(c(1, -0.5), c(-0.5, 1))
z <- sdid_rmvnorm(10000, sigma)
cor(z)
}
}
\keyword{internal}
