% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_engine.R
\name{select_axis_breaks}
\alias{select_axis_breaks}
\title{Select a manageable number of axis breaks}
\usage{
select_axis_breaks(x, max_breaks = 10)
}
\arguments{
\item{x}{Numeric vector of sorted x values}

\item{max_breaks}{Maximum number of breaks}
}
\value{
Numeric vector of break values
}
\description{
Select a manageable number of axis breaks
}
\keyword{internal}
