% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placebo-simulations.R
\name{simulate_dgp}
\alias{simulate_dgp}
\title{Simulates data from DGPs used in the placebo studies in Sections 3 and 5
of the synthetic difference in differences paper. Described there in Section 3.1.1.}
\usage{
simulate_dgp(parameters, N1, T1)
}
\arguments{
\item{parameters, }{a list of dgp parameters (F,M,Sigma,pi) as output by estimate.dgp}

\item{N1, }{a cap on the number of treated units,}

\item{T1, }{the number of treated periods.}
}
\value{
a list with 3 elements: the outcome matrix Y, the number of control units N0, and the number of control periods T0.
The first N0 rows of Y are for units assigned to control, the remaining rows are for units assigned to treatment.
}
\description{
Simulates data from DGPs used in the placebo studies in Sections 3 and 5
of the synthetic difference in differences paper. Described there in Section 3.1.1.
}
