% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{sparsify_function}
\alias{sparsify_function}
\title{A wraper for synthdid_estimate that allows using formula notation and data frames to
estimate synthetic difference in differences.}
\usage{
sparsify_function(v)
}
\arguments{
\item{v}{a vector}
}
\description{
A wraper for synthdid_estimate that allows using formula notation and data frames to
estimate synthetic difference in differences.
}
