% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staggered_methods.R
\name{summary.synthdid_staggered}
\alias{summary.synthdid_staggered}
\title{Summary for synthdid_staggered}
\usage{
\method{summary}{synthdid_staggered}(object, ...)
}
\arguments{
\item{object}{A synthdid_staggered object}

\item{...}{Additional arguments (currently ignored)}
}
\value{
A summary object of class \code{"summary.synthdid_staggered"}
}
\description{
Summary for synthdid_staggered
}
