% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference-solver.R
\name{synthdid.reference}
\alias{synthdid.reference}
\title{Reference implementation of the synthdid estimator}
\usage{
synthdid.reference(
  Y,
  N0,
  T0,
  zeta.omega = ((nrow(Y) - N0) * (ncol(Y) - T0))^(1/4) * sigma.default(Y, N0, T0)
)
}
\arguments{
\item{Y}{Outcome matrix with controls first and pre-treatment periods first.}

\item{N0}{Number of control units.}

\item{T0}{Number of pre-treatment periods.}

\item{zeta.omega}{Ridge penalty for the unit weights \code{omega}.}
}
\value{
Scalar treatment effect estimate.
}
\description{
Computes the synthetic difference-in-differences estimate using convex
optimization for the reference implementation described in the paper.
}
\keyword{internal}
