% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{synthdid_controls}
\alias{synthdid_controls}
\title{Outputs a table of important synthetic controls and their corresponding weights, sorted by weight.}
\usage{
synthdid_controls(
  estimates,
  sort.by = 1,
  mass = SYNTHDID_CONTROLS_MASS_DEFAULT,
  weight.type = "omega"
)
}
\arguments{
\item{estimates, }{a list of estimates output by synthdid_estimate. Or a single estimate.}

\item{sort.by, }{the index of the estimate to sort by. Defaults to 1.}

\item{mass, }{which controls the length of the table. Defaults to 0.9.}

\item{weight.type, }{'omega' for units, 'lambda' for time periods}
}
\value{
A matrix of weights for the top controls/periods, sorted by importance.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{synthdid_controls()} was deprecated in synthdid 2.0.0.
Use the formula interface with \code{\link[=synthdid]{synthdid()}} instead.

The table is truncated to exclude synthetic controls that do not matter for any estimate ---
for each estimate, the truncated controls may have total weight no larger that 1-mass.
}
\keyword{internal}
