% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{synthdid_effect_curve}
\alias{synthdid_effect_curve}
\title{Outputs the effect curve that was averaged to produce our estimate}
\usage{
synthdid_effect_curve(estimate)
}
\arguments{
\item{estimate, }{as output by synthdid_estimate}
}
\value{
A vector of treatment effects for each post-treatment period
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{synthdid_effect_curve()} was deprecated in synthdid 2.0.0.
Use \code{predict(fit, type = "effect")} instead.
}
\keyword{internal}
