% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdid_lib.R
\name{synthdid_impute}
\alias{synthdid_impute}
\title{Imputes control potential outcomes for each element of a treated block, Y[i,j] for i > N_0, j > T_0. 
Each such entry is imputed separately using synthetic diff-in-diff based on control and pretreatment observations in Y.
Use sdid_est instead of this when estimating average treatment effects with multiple missing units,
as it is better to impute the mean of the missing entries than to impute each missing entry individually and average them.}
\usage{
synthdid_impute(Y, N_0, T_0, zeta.lambda = 0,
  zeta.omega = var(as.numeric(Y)), lambda.intercept = FALSE,
  omega.intercept = FALSE, solver = NULL)
}
\arguments{
\item{Y, }{the observation matrix.}

\item{N_0, }{the number of control units. Rows 1-N_0 of Y correspond to the control units.}

\item{T_0, }{the number of pre-treatment time steps. Columns 1-T_0 of Y correspond to pre-treatment time steps.}

\item{zeta.lambda, }{the weight on an L2 penalty on lambda. See (6.1) in the paper. Defaults to zero.}

\item{zeta.omega, }{analogous for omega. Defaults to var(Y).}
}
\value{
A copy of Y with entries Y[i,j] for i > N_0, j > T_0 replaced with imputed values
}
\description{
Imputes control potential outcomes for each element of a treated block, Y[i,j] for i > N_0, j > T_0. 
Each such entry is imputed separately using synthetic diff-in-diff based on control and pretreatment observations in Y.
Use sdid_est instead of this when estimating average treatment effects with multiple missing units,
as it is better to impute the mean of the missing entries than to impute each missing entry individually and average them.
}
