% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{synthdid_placebo}
\alias{synthdid_placebo}
\title{Computes a placebo variant of our estimator using pre-treatment data only}
\usage{
synthdid_placebo(estimate, treated.fraction = NULL)
}
\arguments{
\item{estimate, }{as output by synthdid_estimate}

\item{treated.fraction, }{the fraction of pre-treatment data to use as a placebo treatment period
Defaults to NULL, which indicates that it should be the fraction of post-treatment to pre-treatment data}
}
\value{
A placebo estimate using only pre-treatment data
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{synthdid_placebo()} was deprecated in synthdid 2.0.0.
Use the formula interface with \code{\link[=synthdid]{synthdid()}} instead.
}
\keyword{internal}
