% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{synthdid_placebo_plot}
\alias{synthdid_placebo_plot}
\title{For our estimator and a placebo, plots treated and synthetic control trajectories and overlays a 2x2 diff-in-diff diagram.}
\usage{
synthdid_placebo_plot(estimate, overlay = FALSE, treated.fraction = NULL)
}
\arguments{
\item{estimate, }{as output by synthdid_estimate.}

\item{overlay, }{binary, indicates whether plots should be overlaid or shown in different facets. Defaults to FALSE.}

\item{treated.fraction}{as in synthdid_placebo}
}
\value{
A ggplot2 object showing the estimate and placebo comparison.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{synthdid_placebo_plot()} was deprecated in synthdid 2.0.0.
Use the formula interface with \code{\link[=synthdid]{synthdid()}} instead.

Requires ggplot2
}
\keyword{internal}
