% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{synthdid_se}
\alias{synthdid_se}
\title{Calculate the standard error of a synthetic diff in diff estimate.}
\usage{
synthdid_se(...)
}
\arguments{
\item{...}{Any valid arguments for vcov.synthdid_estimate}
}
\value{
A scalar standard error estimate.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{synthdid_se()} was deprecated in synthdid 2.0.0. Use \code{\link[=vcov]{vcov()}} instead.
}
\keyword{internal}
