% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdid_lib.R
\name{synthdid_simple}
\alias{synthdid_simple}
\title{returns treatment effect curve tau_t for the post-treatment period}
\usage{
synthdid_simple(omega.weight, lambda.weight, Y_00, Y_10, Y_01, Y_11)
}
\description{
returns treatment effect curve tau_t for the post-treatment period
}
