% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{timesteps}
\alias{timesteps}
\title{Get timesteps from panel matrix Y}
\usage{
timesteps(Y)
}
\arguments{
\item{Y}{a matrix}
}
\value{
its column names interpreted as Dates if possible
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{timesteps()} was deprecated in synthdid 2.0.0.
Use the formula interface with \code{\link[=synthdid]{synthdid()}} instead.

timesteps are stored as colnames(Y), but column names cannot be Date objects.
Instead, we use strings. If they are strings convertible to dates, return that
}
\keyword{internal}
