% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{update.synthdid}
\alias{update.synthdid}
\title{Update a synthdid model}
\usage{
\method{update}{synthdid}(object, formula. = NULL, method = NULL, ...)
}
\arguments{
\item{object}{A synthdid object}

\item{formula.}{An optional new formula}

\item{method}{An optional new method}

\item{...}{Additional arguments}
}
\description{
Update and re-fit a synthdid model with different options
}
