% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.synthdid_estimate}
\alias{vcov.synthdid_estimate}
\title{Calculate Variance-Covariance Matrix for a Fitted Model Object}
\usage{
\method{vcov}{synthdid_estimate}(
  object,
  method = c("bootstrap", "jackknife", "placebo"),
  replications = 200,
  ...
)
}
\arguments{
\item{object}{A synthdid model}

\item{method, }{the CI method. The default is bootstrap (warning: this may be slow on large
data sets, the jackknife option is the fastest, with the caveat that it is not recommended
for SC).}

\item{replications, }{the number of bootstrap replications}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Provides variance estimates based on the following three options
\itemize{
\item The bootstrap, Algorithm 2 in Arkhangelsky et al.
\item The jackknife, Algorithm 3 in Arkhangelsky et al.
\item Placebo, Algorithm 4 in Arkhangelsky et al.
}
}
\details{
The jackknife is not recommended for SC, see section 5 in Arkhangelsky et al.
"placebo" is the only option that works for only one treated unit.
}
\references{
Dmitry Arkhangelsky, Susan Athey, David A. Hirshberg, Guido W. Imbens, and Stefan Wager.
"Synthetic Difference in Differences". arXiv preprint arXiv:1812.09970, 2019.
}
