% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placebo-simulations.R
\name{ar2_correlation_matrix}
\alias{ar2_correlation_matrix}
\title{Build the correlation matrix implied by an AR(2) process}
\usage{
ar2_correlation_matrix(ar_coef, T)
}
\arguments{
\item{ar_coef}{Vector of AR(2) coefficients \code{c(phi1, phi2)}.}

\item{T}{Length of the time series.}
}
\value{
Correlation matrix for the implied AR(2) process.
}
\description{
Build the correlation matrix implied by an AR(2) process
}
\keyword{internal}
