% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{did_estimate}
\alias{did_estimate}
\title{synthdid_estimate for diff-in-diff estimates.
Takes all the same parameters, but default, uses constant weights lambda and omega}
\usage{
did_estimate(
  Y,
  N0,
  T0,
  X = array(dim = c(dim(Y), 0)),
  zeta.lambda = 0,
  zeta.omega = sd(apply(Y, 1, diff)),
  lambda.intercept = FALSE,
  omega.intercept = FALSE,
  weights = list(lambda = rep(1/T0, T0), omega = rep(1/N0, N0), vals = NULL),
  min.decrease = 0.001,
  max.iter = 10000
)
}
\description{
synthdid_estimate for diff-in-diff estimates.
Takes all the same parameters, but default, uses constant weights lambda and omega
}
