% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{did_estimate}
\alias{did_estimate}
\title{synthdid_estimate for diff-in-diff estimates.
Takes all the same parameters, but by default, passes options to use the diff-in-diff estimator}
\usage{
did_estimate(Y, N0, T0, ...)
}
\arguments{
\item{Y}{the observation matrix.}

\item{N0}{the number of control units. Rows 1-N0 of Y correspond to the control units.}

\item{T0}{the number of pre-treatment time steps. Columns 1-T0 of Y correspond to pre-treatment time steps.}

\item{...}{additional  options for synthdid_estimate}
}
\value{
an object like that returned by synthdid_estimate
}
\description{
synthdid_estimate for diff-in-diff estimates.
Takes all the same parameters, but by default, passes options to use the diff-in-diff estimator
}
\examples{
\donttest{
# Estimate treatment effect using difference-in-differences
data(california_prop99)
setup <- panel.matrices(california_prop99)

# DID estimate
tau.did <- did_estimate(setup$Y, setup$N0, setup$T0)
print(tau.did)

# Compare all three estimators
tau.sc <- sc_estimate(setup$Y, setup$N0, setup$T0)
tau.sdid <- synthdid_estimate(setup$Y, setup$N0, setup$T0)
estimates <- list(did = tau.did, sc = tau.sc, sdid = tau.sdid)
sapply(estimates, function(x) x)

# Visualize the differences
# synthdid_plot(estimates)
}
}
