% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placebo-simulations.R
\name{estimate_dgp}
\alias{estimate_dgp}
\title{Estimates the DGP parameters used in the placebo studies in Sections 3 and 5
of the synthetic difference in differences paper. Described there in Section 3.1.1.}
\usage{
estimate_dgp(Y, assignment_vector, rank)
}
\arguments{
\item{Y, }{an NxT matrix of outcomes}

\item{assignment_vector, }{an Nx1 vector of treatment assignments}

\item{rank, }{the rank of the estimated signal component L}
}
\value{
a list with elements F, M, Sigma, pi as described in Section 3.1.1
and an element ar_coef with the AR(2) model coefficients underlying the covariance Sigma
}
\description{
Estimates the DGP parameters used in the placebo studies in Sections 3 and 5
of the synthetic difference in differences paper. Described there in Section 3.1.1.
}
\examples{
\donttest{
# Estimate DGP parameters from California Prop 99 data
data(california_prop99)
setup <- panel.matrices(california_prop99)
dgp_params <- estimate_dgp(setup$Y, setup$Y[, ncol(setup$Y)] > 0, rank = 2)

# Use parameters to simulate similar data
simulated <- simulate_dgp(dgp_params, N1 = 1, T1 = 11)
tau.sim <- synthdid_estimate(simulated$Y, simulated$N0, simulated$T0)
}
}
