% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{format.synthdid_estimate}
\alias{format.synthdid_estimate}
\title{Format a synthdid object}
\usage{
\method{format}{synthdid_estimate}(x, ...)
}
\arguments{
\item{x}{The object to format}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Format a synthdid object
}
