% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{jackknife}
\alias{jackknife}
\title{Jackknife standard error of a scalar estimator}
\usage{
jackknife(x, theta)
}
\arguments{
\item{x}{Vector of sample indices.}

\item{theta}{Function returning a scalar estimate when applied to a subset of
\code{x}.}
}
\value{
Numeric jackknife standard error.
}
\description{
Computes the leave-one-out jackknife standard error for a statistic
calculated by \code{theta}.
}
\keyword{internal}
