% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{panel.matrices}
\alias{panel.matrices}
\title{Convert a long (balanced) panel to a wide matrix}
\usage{
panel.matrices(panel, unit = 1, time = 2, outcome = 3, treatment = 4)
}
\arguments{
\item{panel}{A data.frame with columns consisting of units, time, outcome, and treatment indicator.}

\item{unit}{The column number/name corresponding to the unit identifier. Default is 1.}

\item{time}{The column number/name corresponding to the time identifier. Default is 2.}

\item{outcome}{The column number/name corresponding to the outcome identifier. Default is 3.}

\item{treatment}{The column number/name corresponding to the treatment status. Default is 4.}
}
\value{
A list with entries \code{Y}: the data matrix, \code{N0}: the number of control units, \code{T0}:
the number of time periods before treatment.
}
\description{
Converts a data set in panel form to matrix format required by synthdid estimators.
A typical long panel date set look like [unit, time, outcome, treatment]. Synthdid
requires a balanced panel where each unit is observed at all time periods. This function
converts the data set to a num.units * num.time.periods sized matrix where the first N0
rows are treated units, and the first T0 columns are the pre-tretment period.
}
\examples{
\donttest{
# Load tobacco sales in long panel format.
data("california_prop99")
# Transform to N*T matrix format required for synthdid,
# where N is the number of units and T the time periods.
setup <- panel.matrices(california_prop99, unit = 1, time = 2, outcome = 3, treatment = 4)

# Compute synthdid estimate
synthdid_estimate(setup$Y, setup$N0, setup$T0)
}

}
