% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_estimate.R
\name{panel_estimate}
\alias{panel_estimate}
\title{Implement many panel estimators at once. Implements DiD, Synth, Matrix Completion, SDID with and without Intercept and time weights}
\usage{
panel_estimate(
  df,
  unit_id,
  time_id,
  treatment,
  outcome,
  mccores = 8,
  reps = 200,
  methods = c("DID", "Synthetic Control (SC)", "Synthetic DID (SDID)",
    "Matrix Completion"),
  infmethod = c("jackknife", "placebo", "bootstrap")
)
}
\arguments{
\item{df}{data frame (must be a balanced panel)}

\item{unit_id}{identifier for units (column name)}

\item{time_id}{identifier for time periods (column name)}

\item{treatment}{identifier for binary treatment (column name)}

\item{outcome}{identifier for outcome (column name)}

\item{mccores}{number of cores to parallelize inference (whenever possible)}

\item{reps}{number of reps for bootstrap}

\item{methods}{character vector of methods to implement. Must be subset of c("DID", "Synthetic Control (SC)", "Synthetic DID (SDID)", "Time Weighted DID", "SDID (No Intercept)", "SC with FEs (DIFP)", "Matrix Completion", "SC (Regularized)", "DIFP (Regularized)")}

\item{infmethod}{Inference method (must be in c("bootstrap", "jackknife", "placebo"))}
}
\value{
list with matrix with estimates and standard errors and underlying \code{synthdid_estimate} objects, which can individually be plotted / summarized.
}
\description{
Implement many panel estimators at once. Implements DiD, Synth, Matrix Completion, SDID with and without Intercept and time weights
}
\references{
Arkhangelsky et al (2021), "Synthetic Difference in Differences", AER
}
