% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_estimate.R
\name{panel_estimate}
\alias{panel_estimate}
\title{Implement many panel estimators at once. Implements DiD, Synth, Matrix Completion, SDID with and without Intercept and time weights}
\usage{
panel_estimate(df, unit_id, time_id, treatment, outcome)
}
\arguments{
\item{df}{data frame (must be a balanced panel)}

\item{unit_id}{identifier for units (column name)}

\item{time_id}{identifier for time periods (column name)}

\item{treatment}{identifier for binary treatment (column name)}

\item{outcome}{identifier for outcome (column name)}
}
\value{
matrix with estimates and standard errors
}
\description{
Implement many panel estimators at once. Implements DiD, Synth, Matrix Completion, SDID with and without Intercept and time weights
}
\references{
Arkhangelsky et al (2021), "Synthetic Difference in Differences", AER
}
