% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{plot.synthdid_estimate}
\alias{plot.synthdid_estimate}
\title{Plot a synthdid object}
\usage{
\method{plot}{synthdid_estimate}(x, ...)
}
\arguments{
\item{x}{The object to plot}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Plot a synthdid object
}
