% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.synthdid_estimate}
\alias{plot.synthdid_estimate}
\title{Plot a synthdid object}
\usage{
\method{plot}{synthdid_estimate}(x, ...)
}
\arguments{
\item{x}{The object to plot}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A ggplot2 object.
}
\description{
Plot a synthdid object
}
\examples{
\donttest{
data(california_prop99)
setup <- panel.matrices(california_prop99)
tau.hat <- synthdid_estimate(setup$Y, setup$N0, setup$T0)
plot(tau.hat)
}
}
