% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.synthdid_estimate}
\alias{print.summary.synthdid_estimate}
\title{Print summary of synthdid object}
\usage{
\method{print}{summary.synthdid_estimate}(x, ...)
}
\arguments{
\item{x}{A summary.synthdid_estimate object}

\item{...}{Additional arguments}
}
\description{
Print summary of synthdid object
}
