% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{print.synthdid_estimate}
\alias{print.synthdid_estimate}
\title{Print a synthdid object}
\usage{
\method{print}{synthdid_estimate}(x, ...)
}
\arguments{
\item{x}{The object to print}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Print a synthdid object
}
