% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_engine.R
\name{render_weights_plot}
\alias{render_weights_plot}
\title{Render a weights plot}
\usage{
render_weights_plot(spec)
}
\arguments{
\item{spec}{A plot spec from build_plot_spec with type="weights"}
}
\value{
A ggplot2 object
}
\description{
Render a weights plot
}
\keyword{internal}
