% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{sc_estimate}
\alias{sc_estimate}
\title{synthdid_estimate for synthetic control estimates.
Takes all the same parameters, but default, passes options for synthetic control
with no intercept and a penalty term that defaults to the standard deviation of first differences of Y.}
\usage{
sc_estimate(
  Y,
  N0,
  T0,
  X = array(dim = c(dim(Y), 0)),
  zeta.lambda = 0,
  zeta.omega = sd(apply(Y, 1, diff)),
  lambda.intercept = FALSE,
  omega.intercept = FALSE,
  weights = list(lambda = rep(0, T0), omega = NULL, vals = NULL),
  min.step.length = 0
)
}
\description{
synthdid_estimate for synthetic control estimates.
Takes all the same parameters, but default, passes options for synthetic control
with no intercept and a penalty term that defaults to the standard deviation of first differences of Y.
}
