% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdid_lib.R
\name{sc_weight}
\alias{sc_weight}
\title{}
\usage{
sc_weight(
  M,
  target,
  zeta = 1,
  intercept = FALSE,
  constraint = c("simplex", "l1"),
  radius = 2,
  solver = c(CVXR::installed_solvers(), "fw"),
  solver.options = list()
)
}
\arguments{
\item{zeta.}{Defaults to 1.}

\item{intercept.}{Defaults to FALSE}

\item{constraint.}{Defaults to simplex}

\item{radius.}{Defaults to 2, used only when constraint=l1}
}
\value{
gamma, with intercept in attribute
}
\description{

}
