% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdid_lib.R
\name{sc_weight}
\alias{sc_weight}
\title{Computes regularized synthetic control weights
argmin_{gamma}{||M gamma - target||_2^2 / length(target) + zeta * ||gamma||_2^2 : sum gamma = 1, gamma >= 0}.
if intercept = TRUE, these weights gamma solve
argmin_{gamma,c}{||M gamma - target - c||_2^2 / length(target) + zeta * ||gamma||_2^2 : sum gamma = 1, gamma >= 0}
Allowing an intercept may be useful when there are global time trends; see (2.7) in the paper.}
\usage{
sc_weight(M, target, zeta = 1, intercept = FALSE, solver = c("OSQP",
  "ECOS", "SCS"), verbose = FALSE)
}
\arguments{
\item{zeta.}{Defaults to 1.}

\item{intercept.}{Defaults to FALSE}
}
\value{
gamma
}
\description{
Computes regularized synthetic control weights
argmin_{gamma}{||M gamma - target||_2^2 / length(target) + zeta * ||gamma||_2^2 : sum gamma = 1, gamma >= 0}.
if intercept = TRUE, these weights gamma solve
argmin_{gamma,c}{||M gamma - target - c||_2^2 / length(target) + zeta * ||gamma||_2^2 : sum gamma = 1, gamma >= 0}
Allowing an intercept may be useful when there are global time trends; see (2.7) in the paper.
}
