% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{summary.synthdid_estimate}
\alias{summary.synthdid_estimate}
\title{Summarize a synthdid object}
\usage{
\method{summary}{synthdid_estimate}(object, ...)
}
\arguments{
\item{object}{The object to summarize}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Summarize a synthdid object
}
