% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid-package.R
\docType{package}
\name{synthdid-package}
\alias{synthdid}
\alias{synthdid-package}
\title{synthdid: Synthetic Difference-in-Difference Estimation}
\description{
This package implements the synthetic difference in difference estimator (SDID) for the average treatment effect in panel data, as proposed in Arkhangelsky et al (2019). We consider a setting in which we observe a matrix Y = L + tau W + noise where W is a matrix of indicators for treatment. All treated units must begin treatment simultaneously, so W indicates a treated block, i.e. W\link{i,j} = 1 for i > N_0, j > T_0 and is zero otherwise. This applies, in particular, to the case of a single treated unit.

This package is currently in beta and the functionality and interface is subject to change.

Some helpful links for getting started:
\itemize{
\item The \href{https://synth-inference.github.io/synthdid/}{R package documentation} contains usage examples and method reference.
\item The \href{https://synth-inference.github.io/synthdid/articles/more-plotting.html}{online vignettes} contains a gallery of plot examples.
\item For community questions and answers around usage, see \href{https://github.com/synth-inference/synthdid/issues}{Github issues page}.
}
}
\examples{
\donttest{
# Estimate the effect of California Proposition 99 on cigarette consumption
data('california_prop99')
setup = panel.matrices(california_prop99)
tau.hat = synthdid_estimate(setup$Y, setup$N0, setup$T0)
se = sqrt(vcov(tau.hat, method='placebo'))
print(paste0("point estimate: ", round(tau.hat, 2)))
print(paste0("95\% CI for tau: (", round(tau.hat - 1.96 * se, 2), ", ",
                                  round(tau.hat + 1.96 * se, 2), ")"))
}

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/synth-inference/synthdid}
  \item Report bugs at \url{https://github.com/synth-inference/synthdid/issues}
}

}
\keyword{internal}
