% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{synthdid}
\alias{synthdid}
\title{A wrapper function for estimating one or more synthetic diff-in-diffs given a model
formula and a data frame.}
\usage{
synthdid(formula, data, method = synthdid_estimate, ...)
}
\arguments{
\item{formula}{a formula representing the relationship to be estimated. For example,
\code{formula = y ~ d | unit + time} where y is the dependent variable,
d is the time-and-unit-varying treatment variable, and unit and time are
unit and time identifiers, respectively. The unit identifier must precede
the time identifier. Optional control variables \code{x1} and \code{x2} may
be added by appending \code{| ~ x1 + x2} to the end of the formula.
Additionally, you can estimate models of multiple outcomes by expressing the
left hand side as a vector of multiple variables \code{c(y1, y2, y3)}.}

\item{data}{a data.frame object including each variable in formula.}

\item{method}{the estimator to use. One of \code{did_estimate}, \code{sc_estimate}, or \code{synthdid_estimate}.
Default is \code{synthdid_estimate}.}

\item{...}{optional parameters to pass on to synthdid_estimate.}
}
\value{
A list of average treatment effect estimates with 'weights' and 'setup' attached as attributes.
'weights' contains the estimated weights lambda and omega and corresponding intercepts,
as well as regression coefficients beta if X is passed.
'setup' is a list describing the problem passed in: Y, N0, T0, X.
}
\description{
A wrapper function for estimating one or more synthetic diff-in-diffs given a model
formula and a data frame.
}
\examples{
data(CPS)
# No controls
formula_1 <- c(urate, hours) ~ min_wage | state + year
# Controlling for log wage
formula_2 <- c(urate, hours) ~ min_wage | state + year | ~ log_wage
# Estimate models
estimates_nocontrol <- synthdid(formula_1, CPS)
estimates_lwcontrol <- synthdid(formula_2, CPS)

}
