% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{synthdid_controls}
\alias{synthdid_controls}
\title{Outputs a table of important synthetic controls and their corresponding weights, sorted by weight.
The table is truncated to exclude synthetic controls that do not matter for any estimate ---
for each estimate, the truncated controls may have total weight no larger that 1-mass.}
\usage{
synthdid_controls(estimates, sort.by = 1, mass = 0.9, weight.type = "omega")
}
\arguments{
\item{estimates, }{a list of estimates output by synthdid_estimate. Or a single estimate.}

\item{sort.by, }{the index of the estimate to sort by. Defaults to 1.}

\item{mass, }{which controls the length of the table. Defaults to 0.9.}

\item{weight.type, }{'omega' for units, 'lambda' for time periods}
}
\description{
Outputs a table of important synthetic controls and their corresponding weights, sorted by weight.
The table is truncated to exclude synthetic controls that do not matter for any estimate ---
for each estimate, the truncated controls may have total weight no larger that 1-mass.
}
