% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{synthdid_effect_curve}
\alias{synthdid_effect_curve}
\title{Outputs the effect curve that was averaged to produce our estimate}
\usage{
synthdid_effect_curve(estimate)
}
\arguments{
\item{estimate, }{as output by synthdid_estimate}
}
\description{
Outputs the effect curve that was averaged to produce our estimate
}
