% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{synthdid_effect_curve}
\alias{synthdid_effect_curve}
\title{Outputs the effect curve that was averaged to produce our estimate}
\usage{
synthdid_effect_curve(estimate)
}
\arguments{
\item{estimate, }{as output by synthdid_estimate}
}
\value{
A vector of treatment effects for each post-treatment period
}
\description{
Outputs the effect curve that was averaged to produce our estimate
}
\examples{
\donttest{
data(california_prop99)
setup <- panel.matrices(california_prop99)
tau.hat <- synthdid_estimate(setup$Y, setup$N0, setup$T0)

# Get effect curve over time
effect_curve <- synthdid_effect_curve(tau.hat)
plot(effect_curve, type = "l", xlab = "Post-treatment period", ylab = "Effect")
}
}
