% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{synthdid_estimate}
\alias{synthdid_estimate}
\title{Computes synthetic diff-in-diff estimate for an average treatment effect on a treated block.
See Section 4.1 of the paper.}
\usage{
synthdid_estimate(
  Y,
  N0,
  T0,
  X = array(dim = c(dim(Y), 0)),
  zeta.lambda = 0,
  zeta.omega = 0,
  lambda.intercept = TRUE,
  omega.intercept = TRUE,
  weights = NULL,
  start.beta = 0
)
}
\arguments{
\item{Y, }{the observation matrix.}

\item{N0, }{the number of control units. Rows 1-N0 of Y correspond to the control units.}

\item{T0, }{the number of pre-treatment time steps. Columns 1-T0 of Y correspond to pre-treatment time steps.}

\item{X, }{an optional 3-D array of time-varying covariates. Shape should be N X T X C for C covariates.}

\item{zeta.lambda.}{Its square is weight of the ridge penalty relative to MSE. Defaults to 0.}

\item{zeta.omega.}{Analogous for omega. Defaults to 0.}
}
\value{
An average treatment effect estimate, 'weights' and 'setup' attached as attributes.
        Weights contains the estimated weights lambda and omega and corresponding intercepts.
        If covariates X are passedas well as regression coefficients beta if X is passed
        Setup is a list describing the problem passed in: Y, N0, T0, X.
}
\description{
Computes synthetic diff-in-diff estimate for an average treatment effect on a treated block.
See Section 4.1 of the paper.
}
