% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdid_lib.R
\name{synthdid_estimate}
\alias{synthdid_estimate}
\title{Computes synthetic diff-in-diff estimate for an average treatment effect on a treated block. 
See Section 4.1 of the paper. Also computes a jacknife estimate of its standard error.
If fast.var=TRUE, this uses the method described in Section 5 of the paper, in which lambda and omega 
over jackknife replications. If fast.var=FALSE, we re-estimate lambda and omega in each jacknife replication.}
\usage{
synthdid_estimate(Y, N_0, T_0, zeta = var(as.numeric(Y)), fast.var = T,
  lambda.intercept = FALSE, omega.intercept = FALSE, solver = NULL)
}
\arguments{
\item{Y, }{the observation matrix.}

\item{N_0, }{the number of control units. Rows 1-N_0 of Y correspond to the control units.}

\item{T_0, }{the number of pre-treatment time steps. Columns 1-T_0 of Y correspond to pre-treatment time steps.}

\item{zeta, }{the weight on an L2 penalty on the weights. See (6.1) in the paper. Defaults to var(Y).}

\item{fast.var.}{Defaults to TRUE.}
}
\value{
An average treatment effect estimate, with a standard error estimate attached as the attribute 'se'
}
\description{
Computes synthetic diff-in-diff estimate for an average treatment effect on a treated block. 
See Section 4.1 of the paper. Also computes a jacknife estimate of its standard error.
If fast.var=TRUE, this uses the method described in Section 5 of the paper, in which lambda and omega 
over jackknife replications. If fast.var=FALSE, we re-estimate lambda and omega in each jacknife replication.
}
