% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{synthdid_estimate}
\alias{synthdid_estimate}
\title{Computes the synthetic diff-in-diff estimate for an average treatment effect on a treated block.
See Section 4.1 of the paper.}
\usage{
synthdid_estimate(
  Y,
  N0,
  T0,
  X = array(dim = c(dim(Y), 0)),
  noise.level = sd(apply(Y[1:N0, 1:T0], 1, diff)),
  zeta.omega = ((nrow(Y) - N0) * (ncol(Y) - T0))^(1/4) * noise.level,
  zeta.lambda = 1e-06 * noise.level,
  omega.intercept = TRUE,
  lambda.intercept = TRUE,
  weights = list(omega = NULL, lambda = NULL, vals = NULL),
  update.omega = is.null(weights$omega),
  update.lambda = is.null(weights$lambda),
  min.decrease = 0.001 * noise.level,
  max.iter = 10000
)
}
\arguments{
\item{Y}{the observation matrix.}

\item{N0}{the number of control units. Rows 1-N0 of Y correspond to the control units.}

\item{T0}{the number of pre-treatment time steps. Columns 1-T0 of Y correspond to pre-treatment time steps.}

\item{X}{an optional 3-D array of time-varying covariates. Shape should be N X T X C for C covariates.}

\item{noise.level, }{an estimate of the noise standard deviation sigma. Defaults to the standard deviation of first differences of Y.}

\item{zeta.omega}{Analogous for omega. Defaults to (N_tr T_post)^(1/4) * noise.level}

\item{zeta.lambda}{Its square is weight of the ridge penalty relative to MSE. Defaults to an 'infinitesimal' value 1e-6 * noise.level}

\item{omega.intercept}{Binary. Use an intercept when estimating omega.}

\item{lambda.intercept}{Binary. Use an intercept when estimating lambda.}

\item{weights}{a list with fields lambda and omega. If non-null weights$lambda is passed,
we use them instead of estimating lambda weights. Same for weights$omega.}

\item{update.omega}{Analogous.}

\item{update.lambda}{If true, solve for lambda using the passed value of weights$lambda only as an initialization.
If false, use it exactly as passed. Defaults to false if a non-null value of weights$lambda is passed.}

\item{min.decrease}{Tunes a stopping criterion for our weight estimator. Stop after an iteration results in a decrease
in penalized MSE smaller than min.decrease^2.}

\item{max.iter}{A fallback stopping criterion for our weight estimator. Stop after this number of iterations.}
}
\value{
An average treatment effect estimate with 'weights' and 'setup' attached as attributes.
'weights' contains the estimated weights lambda and omega and corresponding intercepts,
as well as regression coefficients beta if X is passed.
'setup' is a list describing the problem passed in: Y, N0, T0, X.
}
\description{
Computes the synthetic diff-in-diff estimate for an average treatment effect on a treated block.
See Section 4.1 of the paper.
}
