% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{synthdid_placebo_plot}
\alias{synthdid_placebo_plot}
\title{For our estimator and a placebo, plots treated and synthetic control trajectories and overlays a 2x2 diff-in-diff diagram.
Requires ggplot2}
\usage{
synthdid_placebo_plot(estimate, overlay = FALSE, treated.fraction = NULL)
}
\arguments{
\item{estimate, }{as output by synthdid_estimate.}

\item{overlay, }{binary, indicates whether plots should be overlaid or shown in different facets. Defaults to FALSE.}

\item{treated.fraction}{as in synthdid_placebo}
}
\description{
For our estimator and a placebo, plots treated and synthetic control trajectories and overlays a 2x2 diff-in-diff diagram.
Requires ggplot2
}
