% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{synthdid_plot}
\alias{synthdid_plot}
\title{Plots treated and synthetic control trajectories and overlays a 2x2 diff-in-diff diagram of our estimator.
In this overlay, the treatment effect is indicated by an arrow.
The weights lambda defining our synthetic pre-treatment time period are plotted below.
Requires ggplot2}
\usage{
synthdid_plot(
  estimates,
  treated.name = "treated",
  control.name = "synthetic control"
)
}
\arguments{
\item{estimates, }{a list of estimates output by synthdid_estimate. Or a single estimate.}

\item{treated.name, }{the name of the treated curve that appears in the legend. Defaults to 'treated'}

\item{control.name, }{the name of the control curve that appears in the legend. Defaults to 'synthetic control'}
}
\description{
Plots treated and synthetic control trajectories and overlays a 2x2 diff-in-diff diagram of our estimator.
In this overlay, the treatment effect is indicated by an arrow.
The weights lambda defining our synthetic pre-treatment time period are plotted below.
Requires ggplot2
}
