% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{synthdid_se}
\alias{synthdid_se}
\title{An estimate of the standard error of our estimator via unit-wise jackknife.
Should not be trusted when the number of treated units is small, and returns NA if there is only one treated unit.
By default, does not recompute the weights omega and lambda. Instead, weights$lambda and weights$beta will be used unchanged and
weights$omega will have jackknifed-out units removed then be renormalized to sum to one.
Pass weights=NULL to recompute weights for each jackknife replication.
Requires bootstrap}
\usage{
synthdid_se(
  estimate,
  method = c("jackknife", "bootstrap", "placebo"),
  weights = attr(estimate, "weights"),
  repetitions = 25,
  treated.fraction = NULL
)
}
\arguments{
\item{estimate, }{as output by synthdid_estimate}

\item{method, }{the CI method}

\item{weights, }{like attr(estimate, 'weights')}

\item{repetitions, }{the number of bootstrap repetitions}

\item{treated.fraction, }{the fraction of pre-treatment data to use as a placebo treatment period
Defaults to NULL, which indicates that it should be the fraction of post-treatment to pre-treatment data}
}
\description{
An estimate of the standard error of our estimator via unit-wise jackknife.
Should not be trusted when the number of treated units is small, and returns NA if there is only one treated unit.
By default, does not recompute the weights omega and lambda. Instead, weights$lambda and weights$beta will be used unchanged and
weights$omega will have jackknifed-out units removed then be renormalized to sum to one.
Pass weights=NULL to recompute weights for each jackknife replication.
Requires bootstrap
}
