% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{synthdid_se}
\alias{synthdid_se}
\title{Calculate the standard error of a synthetic diff in diff estimate. Deprecated. Use vcov.synthdid_estimate.}
\usage{
synthdid_se(...)
}
\description{
Calculate the standard error of a synthetic diff in diff estimate. Deprecated. Use vcov.synthdid_estimate.
}
