% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{synthdid_se}
\alias{synthdid_se}
\title{Calculate standard error for estimator}
\usage{
synthdid_se(
  estimate,
  method = c("bootstrap", "jackknife", "placebo"),
  weights = attr(estimate, "weights"),
  replications = 200
)
}
\arguments{
\item{estimate}{A synthdid model}

\item{method, }{the CI method. The default is bootstrap (warning: this may be slow on large
data sets, the jackknife option is the fastest, with the caveat that it is not recommended
for SC).}

\item{weights, }{like attr(estimate, 'weights')}

\item{replications, }{the number of bootstrap replications}
}
\description{
Provides variance estimates based on the following three options
\itemize{
  \item The bootstrap, Algorithm 2 in Arkhangelsky et al.
  \item The jackknife, Algorithm 3 in Arkhangelsky et al.
  \item Placebo, Algorithm 4 in Arkhangelsky et al.
}
}
\details{
The jackknife is not recommended for SC, see section 5 in Arkhangelsky et al.
}
\references{
Dmitry Arkhangelsky, Susan Athey, David A. Hirshberg, Guido W. Imbens, and Stefan Wager.
 "Synthetic Difference in Differences". arXiv preprint arXiv:1812.09970, 2019.
}
