% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{synthdid_units_plot}
\alias{synthdid_units_plot}
\title{Plots unit by unit difference-in-differences
Requires ggplot2}
\usage{
synthdid_units_plot(
  estimates,
  show.ci = FALSE,
  negligible.threshold = 0.001,
  negligible.alpha = 0.3
)
}
\arguments{
\item{estimates}{as output by synthdid_estimate. Can be a single one or a list of them.}

\item{show.ci}{If TRUE, plots horizontal lines for 95\% CI as well as the point estimate. Defaults to FALSE.}

\item{negligible.threshold}{Unit weight threshold below which units are plotted as small, transparent xs instead of circles. Defaults to .001.}

\item{negligible.alpha}{Determines transparency of those xs.}
}
\description{
Plots unit by unit difference-in-differences
Requires ggplot2
}
