% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{timesteps}
\alias{timesteps}
\title{Get timesteps from panel matrix Y}
\usage{
timesteps(Y)
}
\arguments{
\item{Y}{a matrix}
}
\value{
its column names interpreted as Dates if possible
}
\description{
timesteps are stored as colnames(Y), but column names cannot be Date objects.
Instead, we use strings. If they are strings convertible to dates, return that
}
